refactor: Profile 페이지 리팩토링
