+++ src/pages/LandingPage/_components/Logo.tsx
