+++ src/pages/MainPage/EditChannels/_components/ChannelContainer.tsx
