+++ src/pages/MainPage/EditChannels/_components/ChannelItem.tsx
