+++ src/pages/MainPage/_compoenets/Content/ChannelContainer.tsx
