+++ src/pages/MainPage/_compoenets/Content/ChannelItem.tsx
