+++ src/pages/MainPage/_compoenets/Content/ChannelItem.type.tsx
