+++ src/pages/MainPage/_compoenets/Header/UserProfileImage.type.tsx
