+++ src/pages/MainPage/_compoenets/MusicBar/MusicBar.tsx
