+++ src/pages/MainPage/_compoenets/SiderBar/CategoryName.type.tsx
