+++ src/pages/MainPage/_compoenets/SiderBar/NavigatorContainer.type.tsx
