+++ src/pages/MainPage/_compoenets/SiderBar/NavigatorItem.type.tsx
