+++ src/pages/ProfilePage/ProfilePage.hook.tsx
@@ -1,7 +1,7 @@
 // hooks
 import { useUserStore } from '@/store/useUserStore';
 
-export default function useProfile() {
+export default function useProfilePage() {
   const { user } = useUserStore();
 
   return {
